created Reddit and DateTime components for modularity
