+++ src/components/RedditComponent/RedditComponent.styles.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const StyledRedditComponent = styled.div`
+    width: 40%;
+    height: auto;
+    margin: auto;
+    padding-left: 5%;
+    padding-right: 5%;
+    padding-top: 1%;
+    padding-bottom: 1%;
+    background: #e6e6e6;
+    border-radius: 100px;
+`;
+
